requirements.txt
anyio==3.6.2
APScheduler==3.10.1
backports.zoneinfo==0.2.1
bbcode==1.1.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Deprecated==1.2.13
fastapi==0.95.0
fonttools==4.39.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-resources==5.12.0
kiwisolver==1.4.4
limits==2.8.0
loguru==0.6.0
matplotlib==3.7.1
numpy==1.24.2
opencv-python-headless==4.7.0.72
packaging==22.0
Pillow==9.5.0
pydantic==1.10.7
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
slowapi==0.1.7
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
wrapt==1.15.0
zipp==3.15.0
